Deleted legacy logs rolling upgrade test
