+++ test/framework/src/main/java/org/elasticsearch/index/mapper/NumberFieldBlockLoaderTestCase.java
@@ -11,7 +11,6 @@ package org.elasticsearch.index.mapper;
 
 import org.elasticsearch.datageneration.FieldType;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -49,53 +48,20 @@ public abstract class NumberFieldBlockLoaderTestCase<T extends Number> extends B
         return maybeFoldList(resultList);
     }
 
+    @SuppressWarnings("unchecked")
     private T convert(Object value, T nullValue, Map<String, Object> fieldMapping) {
-        switch (value) {
-            case null -> {
-                return nullValue;
-            }
-
-            // String coercion is true by default
-            case String s -> {
-                if (s.isEmpty()) {
-                    return nullValue;
-                }
-                // Attempt to parse the string as a number. If that fails, the string is malformed, so return null
-                Number parsed = tryParseString(s);
-                if (parsed != null) {
-                    return convert(parsed, fieldMapping);
-                }
-                return null;
-            }
-
-            case Number n -> {
-                return convert(n, fieldMapping);
-            }
-
-            default -> {
-            }
+        if (value == null) {
+            return nullValue;
+        }
+        // String coercion is true by default
+        if (value instanceof String s && s.isEmpty()) {
+            return nullValue;
+        }
+        if (value instanceof Number n) {
+            return convert(n, fieldMapping);
         }
 
         // Malformed values are excluded
-        return null;
-    }
-
-    /**
-     * Tries to parse a string as a number, matching the behavior of UnsignedLongFieldMapper.parseUnsignedLong().
-     * Returns null if the string cannot be parsed as a valid number.
-     */
-    private Number tryParseString(String s) {
-        try {
-            return Long.parseUnsignedLong(s);
-        } catch (NumberFormatException ex1) {
-            try {
-                BigDecimal bd = new BigDecimal(s);
-                return bd.toBigIntegerExact();
-            } catch (NumberFormatException | ArithmeticException ex2) {
-                // not a valid number
-            }
-        }
-
         return null;
     }
